auth/requirements.txt
grpcio
grpcio-tools
protobuf
SQLAlchemy
SQLAlchemy-Utils
cryptography
python-dotenv
bcrypt
pyjwt